backend/pyproject.toml
[tool.poetry]
name = "virtual-closet-backend"
version = "1.0.0"
description = "Backend API for Virtual Closet - AI-powered wardrobe management and styling"
authors = ["Virtual Closet Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pydantic = "^2.6.1"
pydantic-settings = "^2.2.1"
pillow = "^10.2.0"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"
torch = "^2.2.0"
torchvision = "^0.17.0"
aiofiles = "^23.2.1"
python-magic = "^0.4.27"
httpx = "^0.27.0"
requests = "^2.31.0"
tqdm = "^4.66.2"
scikit-learn = "^1.4.1"
diffusers = "^0.27.0"
transformers = "^4.38.0"
accelerate = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true